Added Naloxone and a clunky script
shift_levels.py is meant to be run after using convert_levels.py, and is really pretty sucky. But it works. Basically you just pass it a path to a level file, an x-coordinate, and a z-coordinate. The x/z coordinates should be the actual center of the level. All it does is shift all the map objects around so that the center is at (0,0) as it should be.